refactor: move components to a separate dir
